feat(nu-prompt): add space again before branch symbol
nice job guys at ghostty, let's forget to test a change
that literally redoes rendering glyphs in a dev-geared
terminal (literally half the user base prob has glyphs
somewhere in their terminal). no, seriously, who the fuck
thought it was a good idea for glyphs to take up two spaces
when there's a whitespace next to it, and who the fuck
thought to squish them so much. really, fuck you.